fixture: keep Envira HTML polish only
