+++ src/components/preverif/FixtureBackedVm0007ReportView.tsx
@@ -1,7 +1,5 @@
 import type { ReactNode } from "react";
 import {
-  getPriorityClientActionRows,
-  groupEvidenceMapRowsByStatus,
   type Vm0007EvidenceMapRow,
   type Vm0007FixtureBackedReport,
   type Vm0007FixtureBackedStatus,
@@ -11,6 +9,41 @@ type FixtureBackedVm0007ReportViewProps = {
   report: Vm0007FixtureBackedReport;
   pdfDownloadHref?: string | null;
 };
+
+const VM0007_FIXTURE_BACKED_STATUS_ORDER: Vm0007FixtureBackedStatus[] = [
+  "MISSING",
+  "UNCLEAR",
+  "FOUND",
+  "N/A",
+];
+
+function compareRuleIds(left: string, right: string): number {
+  return left.localeCompare(right, undefined, { numeric: true });
+}
+
+function sortEvidenceMapRows(rows: Vm0007EvidenceMapRow[]): Vm0007EvidenceMapRow[] {
+  const order = new Map(VM0007_FIXTURE_BACKED_STATUS_ORDER.map((status, index) => [status, index]));
+  return rows.slice().sort((left, right) => {
+    const statusDiff = (order.get(left.status) ?? 99) - (order.get(right.status) ?? 99);
+    if (statusDiff !== 0) return statusDiff;
+    return compareRuleIds(left.ruleId, right.ruleId);
+  });
+}
+
+function groupEvidenceMapRowsByStatus(rows: Vm0007EvidenceMapRow[]): Array<{
+  status: Vm0007FixtureBackedStatus;
+  rows: Vm0007EvidenceMapRow[];
+}> {
+  const grouped = VM0007_FIXTURE_BACKED_STATUS_ORDER.map((status) => ({
+    status,
+    rows: sortEvidenceMapRows(rows).filter((row) => row.status === status),
+  }));
+  return grouped.filter((group) => group.rows.length > 0);
+}
+
+function getPriorityClientActionRows(rows: Vm0007EvidenceMapRow[]): Vm0007EvidenceMapRow[] {
+  return sortEvidenceMapRows(rows).filter((row) => row.status === "MISSING" || row.status === "UNCLEAR");
+}
 
 function statusTone(status: Vm0007FixtureBackedStatus): string {
   if (status === "FOUND") return "border-emerald-200 bg-emerald-50 text-emerald-900";
